Update README.md
Hangman game is a simple game made with Python.The Hangman Game is a word-guessing challenge where players guess letters to figure out a hidden word within a limited number of attempts. With each incorrect guess, part of a "hangman" figure is drawn. The player wins by guessing the word before the figure is fully drawn. It improves vocabulary and offers fun gameplay that’s easy to extend or modify. The game is ideal for both educational and recreational purposes.